requirements.txt
# LangGraph and LangChain
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0

# OpenAI
openai>=1.0.0

# Google APIs
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.0.0
gspread>=6.0.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
